Exclude .less/.cson/.coffee files at root of src
Previously the rsync exclude rules required files to be two-levels
deep to be excluded which would keep the window-bootstrap.coffee in
the bundle since it was at the root of the src folder.

Now if cson, less, or coffee files end up in src/ they will be
correctly excluded from the bundle when synced.