Don't export Model superclass in top-level atom module
Previously, the semantics of telepath's Model class were coupled to the
wrapped documents, so it made sense to export the Model. But now that
Model is much simpler, it makes more sense to allow individual packages
to rely on their own version of Model by requiring theorist if they want
to so we can make changes to the version of model used by core without
breaking packages.